+++ SSEngineDefault/SSEngineDefault/Public/CommonSerializer/DefaultTypeSerializsers.h
@@ -21,28 +21,28 @@ SSENGINEDEFAULT_MODULE int64 AppendData(
 /// return when Written Byte Size.
 /// if something went wrong, return 0
 /// </returns>
-SSENGINEDEFAULT_MODULE int32 FillMemoryFromData(
+SSENGINEDEFAULT_MODULE int64 FillMemoryFromData(
 	void* Dest,
-	int CopySize,
+	int64 CopySize,
 	const SS::PooledList<byte>& FromData,
-	int FromOffset);
+	int64 FromOffset);
 
 
 
-SSENGINEDEFAULT_MODULE int32 AppendDataFromStrings(
+SSENGINEDEFAULT_MODULE int64 AppendDataFromStrings(
 	SS::PooledList<byte>& Data,
 	const SS::PooledList<SS::StringW>& Strings);
 
-SSENGINEDEFAULT_MODULE int32 FillStringFromData(
+SSENGINEDEFAULT_MODULE int64 FillStringFromData(
 	SS::PooledList<SS::StringW>& Strings,
 	const SS::PooledList<byte>& Data,
-	int Offset = 0);
+	int64 Offset = 0);
 
-SSENGINEDEFAULT_MODULE int32 AppendDataFromHashers(
+SSENGINEDEFAULT_MODULE int64 AppendDataFromHashers(
 	SS::PooledList<byte>& Data,
 	const SS::PooledList<SS::SHasherW>& Hashers);
 
-SSENGINEDEFAULT_MODULE int32 FillHashersFromData(
+SSENGINEDEFAULT_MODULE int64 FillHashersFromData(
 	SS::PooledList<SS::SHasherW>& Hashers,
 	const SS::PooledList<byte>& Data,
-	int Offset = 0);
+	int64 Offset = 0);
